backend/_archive/app/ai/summary.py
from sqlalchemy.orm import Session
from google import genai
from app.core.config import settings

def generate_health_summary(db: Session, start_date: str, end_date: str):
    \"\"\"
    Stub for using Gemini to generate a health trends summary.
    Workflow:
    1. Query DailyHealthSummary and LabResult for the requested date range.
    2. Format the retrieved data into a markdown string.
    3. Send the formatted data to Gemini using the google-genai SDK with a prompt inquiring about trends.
    4. Save the returned string into the AISummary table.
    \"\"\"
    if not settings.GEMINI_API_KEY:
         print("Warning: GEMINI_API_KEY not set")
    else:
         client = genai.Client(api_key=settings.GEMINI_API_KEY)
         # model = "gemini-2.5-flash-8b" or other preferred
    
    print(f"Generating summary from {start_date} to {end_date} (Not Implemented)")
    pass
