backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
lxml==5.3.0
pandas==2.2.3
pdfplumber==0.11.4
google-genai==0.3.0
python-dotenv==1.0.1
openpyxl==3.1.5
